Skip cleanup of named disposables
This error was uncaught because it was introduced at different commits
and different merge requests, and each one was fixing an issue important
to them, but the last one enabled "force", which should not clean up the
named disposable if it was not for the first commit.

- f1cbf2159
- a82f4f2af

The "force" parameter is not a good name for what it does. Remove it.
And is not needed if the methods and events are ordered correctly and
shared state runs within a lock.

Introduce tests for cases where cleanup is called implicitly or
explicitly:

- Implicit:
  - Failed startup
  - Shutdown/Kill
- Explicit:
  - Cleanup when not running
  - Cleanup when running

Fixes: https://github.com/QubesOS/qubes-issues/issues/11042
Fixes: https://github.com/QubesOS/qubes-issues/issues/10928

+++ qubes/vm/dispvm.py
@@ -699,8 +699,10 @@ class DispVM(qubes.vm.qubesvm.QubesVM):
     @qubes.events.handler("domain-shutdown")
     async def on_domain_shutdown(self, _event, **_kwargs) -> None:
         """
-        Do auto cleanup if enabled.
+        Remove domain if appropriate.
         """
+        if not self.auto_cleanup:
+            return
         await self._delete_domain()
 
     @qubes.events.handler("domain-remove-from-disk")
@@ -955,46 +957,33 @@ class DispVM(qubes.vm.qubesvm.QubesVM):
             )
             self.template.remove_preload_from_list([self.name])
 
-    async def _delete_domain(self, force: bool = False) -> None:
+    async def _delete_domain(self) -> None:
         """
-        Delete domain if auto cleanup is enabled.
-
-        :param bool force: Auto clean up even if property is disabled
+        Delete the disposable qube.
         """
-        if not self.auto_cleanup and not force:
-            return
         self._preload_cleanup()
+        if not hasattr(self, "app"):
+            return
         if self not in self.app.domains:
             return
         del self.app.domains[self]
         await self.remove_from_disk()
         self.app.save()
 
-    async def cleanup(self, force: bool = False) -> None:
+    async def cleanup(self) -> None:
         """
         Clean up after the disposable.
 
         This stops the disposable qube and removes it from the store.
         This method modifies :file:`qubes.xml` file.
-
-        :param bool force: Auto clean up if property is enabled and domain \
-                is not running, should be used in special circumstances only \
-                as the sole purpose of this option is because using it may not \
-                be reliable.
         """
         if self not in self.app.domains:
             return
-        running = True
         try:
             await self.kill()
         except qubes.exc.QubesVMNotStartedError:
-            running = False
-        # Full cleanup will be done automatically if event 'domain-shutdown' is
-        # triggered and "auto_cleanup=True".
-        if not self.auto_cleanup or (
-            force and not running and self.auto_cleanup
-        ):
-            await self._delete_domain(force=force)
+            pass
+        await self._delete_domain()
 
     async def start(self, **kwargs):
         """
@@ -1010,8 +999,13 @@ class DispVM(qubes.vm.qubesvm.QubesVM):
                 )
             await super().start(**kwargs)
         except:
-            # Cleanup also on failed startup
-            await self.cleanup()
+            if self not in self.app.domains:
+                return
+            try:
+                await self.kill()
+            except qubes.exc.QubesVMNotStartedError:
+                if self.auto_cleanup:
+                    await self._delete_domain()
             raise
 
     def create_qdb_entries(self) -> None:
